+++ app/src/main/java/com/example/myapplication/V/Activity/ScaleActivity.java
@@ -37,7 +37,7 @@ public class ScaleActivity extends AppCompatActivity {
             recyclerView.setAdapter(itemAdapter);
 
 
-            list.add(new ItemScale("Horizontal Scale",counter.getHorizontal_scale()+" sec per page"));
+            list.add(new ItemScale("Horizontal Scale",(int)counter.getHorizontal_scale()+" sec per page"));
             list.add(new ItemScale("EEG Sensivity","70µV per page"));
